Se cambió el tipo de datos de los id de los modelos de Documento y DocumentoServicio

+++ src/main/java/com/smty/ApiServiciosProfesionales/Models/Documento.java
@@ -12,7 +12,7 @@ public class Documento {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id_documento")
-    private int idDocumento;
+    private Long idDocumento;
 
     @NotEmpty
 	@Column(name="url", length = 100)
